added empty input field func

+++ includes/deleteItem.inc.php
@@ -7,6 +7,11 @@ if(isset($_POST["deleteItem"])){
     require_once '../includes/dbConnection.inc.php';
     require_once '../includes/functions.inc.php';
 
+    if(emptyDeleteForm($name)){
+        header("location:../deleteItemPage.php?error=emptyInputAtDeleteForm");
+        exit();
+    }
+
     deleteItem($conn, $name);
 }
 else{
